feat: add tracked macro functionality

+++ src/main/kotlin/me/moeszyslak/polly/data/Permissions.kt
@@ -1,14 +1,14 @@
 package me.moeszyslak.polly.data
 
 import dev.kord.core.any
-import me.jakejmattson.discordkt.api.dsl.PermissionContext
-import me.jakejmattson.discordkt.api.dsl.PermissionSet
+import me.jakejmattson.discordkt.dsl.PermissionContext
+import me.jakejmattson.discordkt.dsl.PermissionSet
 
 @Suppress("unused")
 enum class Permissions : PermissionSet {
     BOT_OWNER {
         override suspend fun hasPermission(context: PermissionContext) =
-            context.discord.getInjectionObjects<Configuration>().botOwner == context.user.id.value
+            context.discord.getInjectionObjects<Configuration>().botOwner == context.user.id.value.toLong()
     },
     GUILD_OWNER {
         override suspend fun hasPermission(context: PermissionContext) =
